requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
python-dotenv==1.0.1
openpyxl==3.1.2
Pillow==10.2.0
email-validator==2.1.0.post1
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
Flask-Mail
Flask-WeasyPrint 
psutil==5.9.8 
pyjwt
bcrypt 
gunicorn
flask-limiter 
  reportlab
flasgger>=0.9.7